`rffs_root.asm`
Created the initial `rffs_root` example, this code provides a sample as to how
the bootsector of any system that intends to use rffs should appear, i.e. what
data it should contain and a general overview of how a bootsector is interacted
with by the BIOS. More detailed information about the document can be found in
`docs/RedFox32_Filesyste_Specification.md`.